fix: split OperandField constants/utils to operandValidation.ts (#PAT-162)
ESLint react-refresh/only-export-components requires component files to
only export components. OperandField.tsx mixed component + constants
(NUMERIC_LITERAL_RE / UCUM_UNIT_RE / quantityToCql / OperandMode), which
triggered a warning, which under --max-warnings 0 became a CI failure.

Move the shared constants/utils to a new file operandValidation.ts.
OperandField.tsx now only re-exports the component (default export). Both
consumers (ArithmeticElement, ArithmeticUnaryElement) updated to import
from operandValidation.

17/17 local tests still pass; lint clean.
