+++ frontend/src/components/authoring/base-elements/ArithmeticUnaryElement.tsx
@@ -6,7 +6,8 @@ import { alpha } from '@mui/material/styles'
 import DeleteIcon from '@mui/icons-material/Delete'
 import type { BaseElement } from '../../../types/authoring'
 import { escapeCqlIdentifier } from '../../../utils/cqlString'
-import OperandField, { NUMERIC_LITERAL_RE, quantityToCql, type OperandMode } from './OperandField'
+import OperandField from './OperandField'
+import { NUMERIC_LITERAL_RE, quantityToCql, type OperandMode } from './operandValidation'
 
 // PAT-162: unary CQL functions taking one Decimal/Integer/Quantity argument.
 // Kept alphabetical for predictable dropdown order; mirrors the backend
